feat: seller role 필요한 쿠폰 관련 API 따로 분리
